♻️ use MessageForm in ChatScreen
